visual adjustments, error views and other minor changes

+++ data/src/main/java/com/beetrack/bitcoinwallet/data/remote/data/AddressBalanceData.kt
@@ -5,29 +5,29 @@ import com.google.gson.annotations.SerializedName
 data class AddressBalanceData(
 
 	@field:SerializedName("final_balance")
-	val finalBalance: Int? = null,
+	val finalBalance: Long? = null,
 
 	@field:SerializedName("unconfirmed_n_tx")
-	val unconfirmedNTx: Int? = null,
+	val unconfirmedNTx: Long? = null,
 
 	@field:SerializedName("address")
 	val address: String,
 
 	@field:SerializedName("balance")
-	val balance: Int? = null,
+	val balance: Long? = null,
 
 	@field:SerializedName("final_n_tx")
-	val finalNTx: Int? = null,
+	val finalNTx: Long? = null,
 
 	@field:SerializedName("n_tx")
-	val nTx: Int? = null,
+	val nTx: Long? = null,
 
 	@field:SerializedName("total_sent")
-	val totalSent: Int? = null,
+	val totalSent: Long? = null,
 
 	@field:SerializedName("unconfirmed_balance")
-	val unconfirmedBalance: Int? = null,
+	val unconfirmedBalance: Long? = null,
 
 	@field:SerializedName("total_received")
-	val totalReceived: Int? = null,
+	val totalReceived: Long? = null,
 )
